+++ R/long_to_wide_converter.R
@@ -1,6 +1,5 @@
-#' @title Converts long-format dataframe to wide-format dataframe
+#' @title Converts dataframe from long/tidy to wide format with `NA`s removed
 #' @name long_to_wide_converter
-#' @author \href{https://github.com/IndrajeetPatil}{Indrajeet Patil}
 #' @description This conversion is helpful mostly for repeated measures design.
 #'
 #' @param data A dataframe (or a tibble) from which variables specified are to
@@ -11,11 +10,14 @@
 #' @param paired Logical that decides whether the experimental design is
 #'   repeated measures/within-subjects or between-subjects. The default is
 #'   `FALSE`.
+#' @param spread Logical that decides whether the dataframe needs to be
+#'   converted from long/tidy to wide (default: `TRUE`), or the data needs to be
+#'   returned as it is but with the `NA`s removed.
 #' @param ... Currently ignored.
 #'
 #' @importFrom rlang :=
 #' @importFrom dplyr n row_number select mutate mutate_at group_by ungroup
-#' @importFrom tidyr spread
+#' @importFrom tidyr pivot_longer
 #'
 #' @return A dataframe in the wide (or Cartesian) format.
 #'
@@ -35,47 +37,36 @@ long_to_wide_converter <- function(data,
                                    x,
                                    y,
                                    paired = TRUE,
+                                   spread = TRUE,
                                    ...) {
   # make sure both quoted and unquoted arguments are allowed
   c(x, y) %<-% c(rlang::ensym(x), rlang::ensym(y))
 
-  # creating a dataframe
+  # initial cleanup
   data %<>%
     dplyr::select(.data = ., {{ x }}, {{ y }}) %>%
     dplyr::mutate(.data = ., {{ x }} := droplevels(as.factor({{ x }}))) %>%
-    tibble::as_tibble(x = .)
-
-  # figuring out number of levels in the grouping factor
-  x_n_levels <- nlevels(data %>% dplyr::pull({{ x }}))[[1]]
-
-  # wide format
-  data_wide <-
-    data %>%
-    dplyr::filter(.data = ., !is.na({{ x }})) %>%
+    as_tibble(.) %>%
     dplyr::group_by(.data = ., {{ x }}) %>%
     dplyr::mutate(.data = ., rowid = dplyr::row_number()) %>%
-    dplyr::ungroup(x = .) %>%
-    dplyr::filter(.data = ., !is.na({{ y }}))
+    dplyr::ungroup(.)
 
-  # clean up for repeated measures design
+  # NA removal
   if (isTRUE(paired)) {
-    data_wide %<>%
-      dplyr::group_by(.data = ., rowid) %>%
-      dplyr::mutate(.data = ., n = dplyr::n()) %>%
-      dplyr::ungroup(x = .) %>%
-      dplyr::filter(.data = ., n == x_n_levels) %>%
-      dplyr::select(.data = ., {{ x }}, {{ y }}, rowid)
+    data %<>% dplyr::anti_join(x = ., y = dplyr::filter(., is.na({{ y }})), by = "rowid")
+  } else {
+    data %<>% tidyr::drop_na(.)
   }
 
-  # spreading the columns of interest
-  data_wide %<>%
-    tidyr::spread(
-      data = .,
-      key = {{ x }},
-      value = {{ y }},
-      convert = TRUE
-    )
-
-  # return the dataframe in wide format
-  return(data_wide)
+  # convert to wide?
+  if (isTRUE(spread)) {
+    data %>%
+      tidyr::pivot_wider(
+        data = .,
+        names_from = {{ x }},
+        values_from = {{ y }}
+      )
+  } else {
+    data
+  }
 }