+++ R/stats_type_switch.R
@@ -1,9 +1,9 @@
 #' @name stats_type_switch
 #' @title Switch type of statistics.
-#' @description Relevant mostly for `ggstatsplot` and `statsExpressions`, where
-#'   there are four types of statistics are supported: parametric,
-#'   non-parametric, robust, and Bayesian. This switch function converts strings
-#'   entered by users to a common pattern.
+#' @description Relevant mostly for `ggstatsplot` and `statsExpressions`
+#'   packages, where there are four types of statistics are supported:
+#'   parametric, non-parametric, robust, and Bayesian. This switch function
+#'   converts strings entered by users to a common pattern.
 #'
 #' @param type Character string describing the type of statistics.
 #'
@@ -17,7 +17,7 @@
 stats_type_switch <- function(type) {
   dplyr::case_when(
     grepl("^p", type, TRUE) ~ "parametric",
-    grepl("^n|^s", type, TRUE) ~ "nonparametric", # s is for spearman's rho
+    grepl("^n|^s", type, TRUE) ~ "nonparametric", # s is for Spearman's rho
     grepl("^r", type, TRUE) ~ "robust",
     grepl("^b", type, TRUE) ~ "bayes",
     TRUE ~ "parametric"