+++ NEWS.md
@@ -1,3 +1,11 @@
+# ipmisc 3.1.0
+
+  - Minor refactoring of `long_to_wide_converter`. It gains `spread` argument.
+    When set to `FALSE`, the original dataframe will be returned with the `NA`'s
+    removed.
+    
+  - New helper function `p_value_formatter`.
+
 # ipmisc 3.0.0
 
 BREAKING CHANGES